Move ParameterManager
